# doc/swagger-daily.yaml
swagger: '2.0'
info:
  description: >-
    交接班管理、日报管理、月报管理、专刊管理等.
  version: 1.0.0
  title: daily工作台日常管理
host: 127.0.0.1:8000
basePath: /cdp-daily/v1
...
schemes:
  - http
paths:
...
 /issues/:
   post:
     tags:
       - 专刊管理issue
     summary: '新增专刊信息（后台同步调用大数据平台的接口存入大数据平台）'
     operationId: addIssue
     parameters:
       - name: issue
         in: body
         required: true
         schema:
           $ref: '#/definitions/IssueVo'
     responses:
       200:
         description: ok
         schema:
           $ref: '#/definitions/ReturnResult'
 /issues/{id}:
   put:
     tags:
       - 专刊管理issue
     summary: '修改专刊信息'
     operationId: updateIssue
     parameters:
       - $ref: '#/parameters/id-p'
       - name: issue
         in: body
         required: true
         schema:
           $ref: '#/definitions/IssueVo'
     responses:
       200:
         description: ok
         schema:
           $ref: '#/definitions/ReturnResult'
   get:
     tags:
       - 专刊管理issue
     summary: '获取单条专刊信息'
     operationId: findIssueById
     parameters:
       - $ref: '#/parameters/id-p'
     responses:
       200:
         description: ok
         schema:
           type: object
           allOf:
             - $ref: '#/definitions/ReturnResult'
           properties:
             data:
               $ref: '#/definitions/Issue'
   delete:
     tags:
       - 专刊管理issue
     summary: '删除专刊信息'
     operationId: deleteIssue
     parameters:
       - $ref: '#/parameters/id-p'
     responses:
       200:
         description: ok
         schema:
           $ref: '#/definitions/ReturnResult'
 /issues/search:
    post:
        tags:
          - 专刊管理issue
        summary: '根据条件查询专刊信息列表-分页（与大数据平台对接进行标题和附件内容的检索）'
        operationId: findIssues
        parameters:
          - name: issueVO
            in: body
            required: true
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: 专刊标题
                page:
                  type: integer
                size:
                  type: integer
        responses:
          200:
            description: OK
            schema:
              $ref: '#/definitions/IssueList'
 /issues/publish:
    put:
        tags:
          - 专刊管理issue
        summary: '专刊信息发布，后台将该专刊状态置为已发布状态'
        operationId: publishIssue
        parameters:
          - name: issueId
            in: query
            description: 专刊ID
            type: string
        responses:
          200:
            description: OK
            schema:
              $ref: '#/definitions/ReturnResult'

################ definitions#################
definitions:
...
 Issue:
   type: object
   properties:
     id:
       type: string
     title:
       type: string
       description: 列表显示
     createBy:
       type: string
       description: 列表显示
     cteateTime:
       type: string
       format: date-time
       description: 列表显示
     status:
       type: string
       description: 状态（0未发布，1已发布）
       enum:
         - 0
         - 1
 IssueList:
   type: object
   allOf:
     - $ref: '#/definitions/ReturnResult'
   properties:
     data:
       type: object
       properties:
         totalElements:
           type: integer
           description: 记录总数
         totalPages:
           type: integer
           description: 总页数
         first:
           type: boolean
           description: 是否第一页
         last:
           type: boolean
           description: 是否最后一页
         number:
           type: integer
           description: 当前页数，页码从0开始
         size:
           type: integer
           description: 每页记录数
         sort:
           type: string
         numberOfElements:
           type: integer
           description: 返回记录条数
         content:
           type: array
           items:
             $ref: '#/definitions/Issue'
 IssueVo:
   type: object
   properties:
     issue:
       $ref: '#/definitions/Issue'
     fileIds:
       type: array
       items:
         type: string
         description: 专刊相关附件ID数组，前端传入
     fileDeleteIds:
       type: array
       items:
         type: string
         description: 删除的附件ID数组，前端传入


##### common model####
 ReturnResult:
    type: object
    properties:
      code:
        type: integer
        description: >-
          2000  操作成功
          4000 操作失败
          4001 验证码出错
          4002 用户名或密码错误
          5000 后台逻辑错误
          5001 令牌失效或过期
          5009 未知错误
      msg:
        type: string

############################ parameters#####################
parameters:
  id-p:
    name: id
    in: path
    required: true
    description: 主键ID
    type: string